+++ tests/results.py
@@ -128,6 +128,7 @@ class AddressTable(TableProtocol):
 
     indexes: tuple[tuple[str, ...], ...] = ()
     unique_indexes: tuple[tuple[str, ...], ...] = ()
+    without_rowid: bool = False
     relations: tuple[TableRelation, ...] = (
         TableRelation(attribute="user", remote_table=lambda: UserTable, many=False, mapping={"user_id": "id"}),
     )
@@ -294,6 +295,7 @@ class BirthDayTable(TableProtocol):
 
     indexes: tuple[tuple[str, ...], ...] = ()
     unique_indexes: tuple[tuple[str, ...], ...] = ()
+    without_rowid: bool = False
     relations: tuple[TableRelation, ...] = (
         TableRelation(attribute="user", remote_table=lambda: UserTable, many=False, mapping={"user_id": "id"}),
     )
@@ -457,6 +459,7 @@ class BookTable(TableProtocol):
 
     indexes: tuple[tuple[str, ...], ...] = (('name',),)
     unique_indexes: tuple[tuple[str, ...], ...] = ()
+    without_rowid: bool = False
     relations: tuple[TableRelation, ...] = (
         TableRelation(attribute="users", remote_table=lambda: UserBookTable, many=True, mapping={"id": "book_id"}),
     )
@@ -650,6 +653,7 @@ class CompositeTable(TableProtocol):
 
     indexes: tuple[tuple[str, ...], ...] = ()
     unique_indexes: tuple[tuple[str, ...], ...] = (('uniq1', 'uniq2'), ('uniq3',),)
+    without_rowid: bool = False
     relations: tuple[TableRelation, ...] = ()
 
     def primary_values(self, instance: Composite) -> tuple[int, int]:
@@ -884,6 +888,7 @@ class UserTable(TableProtocol):
 
     indexes: tuple[tuple[str, ...], ...] = (('name',), ('name', 'email'), ('last_login',),)
     unique_indexes: tuple[tuple[str, ...], ...] = (('name', 'email'),)
+    without_rowid: bool = False
     relations: tuple[TableRelation, ...] = (
         TableRelation(attribute="addresses", remote_table=lambda: AddressTable, many=True, mapping={"id": "user_id"}),
         TableRelation(attribute="birthday", remote_table=lambda: BirthDayTable, many=False, mapping={"id": "user_id"}),
@@ -1086,6 +1091,7 @@ class UserBookTable(TableProtocol):
 
     indexes: tuple[tuple[str, ...], ...] = (('created_at',),)
     unique_indexes: tuple[tuple[str, ...], ...] = ()
+    without_rowid: bool = False
     relations: tuple[TableRelation, ...] = (
         TableRelation(attribute="user", remote_table=lambda: UserTable, many=False, mapping={"user_id": "id"}),
         TableRelation(attribute="book", remote_table=lambda: BookTable, many=False, mapping={"book_id": "id"}),